feat: add Author Applications management tab with filtering and action dialogs
feat: create MetricCard component for displaying key metrics in overview

feat: implement OverviewTab for displaying dashboard statistics and recent activities

feat: add StatusBadge component for displaying system health statuses

feat: create SystemHealthTab to monitor database, cache, and storage health
